refactor: Mock data 분리
