+++ src/mocks/handlers.ts
@@ -1,26 +1,13 @@
-import { HTTPSTATUS } from "api/types";
-import { rest } from "msw";
-import { successfulAllRegionsData, successfulCategoriesData } from "./data";
 import authHandlers from "./handlers/authHandlers";
+import categoriesHandlers from "./handlers/categoriesHandlers";
 import productItemHandlers from "./handlers/productItemHandlers";
+import regionsHandlers from "./handlers/regionsHandlers";
 import userHandlers from "./handlers/userHandlers";
 
 export default [
   ...authHandlers,
   ...userHandlers,
   ...productItemHandlers,
-
-  rest.get("/api/categories", async (_, res, ctx) => {
-    return res(
-      ctx.status(HTTPSTATUS.success),
-      ctx.json(successfulCategoriesData)
-    );
-  }),
-
-  rest.get("/api/regions", async (_, res, ctx) => {
-    return res(
-      ctx.status(HTTPSTATUS.success),
-      ctx.json(successfulAllRegionsData)
-    );
-  }),
+  ...categoriesHandlers,
+  ...regionsHandlers,
 ];
